# backend/clazzziks/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLAZZZIKS API",
    "version": "0.1.0",
    "description": "The shared contract between the CLAZZZIKS backend (FastAPI) and frontend (React).\nThis document is the single source of truth for the `/api` surface. The backend\nserves it at `GET /api/openapi.json` and its responses are validated against the\nschemas here by `backend/tests/test_contract.py`. The frontend client\n(`frontend/src/api.js`) builds against these same shapes."
  },
  "servers": [
    { "url": "/api", "description": "Mounted under /api (Vite proxies this to the FastAPI server in dev)." }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Liveness probe.",
        "responses": {
          "200": {
            "description": "Service is up.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Health" }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "operationId": "getOpenapi",
        "summary": "This contract document.",
        "responses": {
          "200": {
            "description": "The OpenAPI contract.",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/formats": {
      "get": {
        "operationId": "getFormats",
        "summary": "Supported formats and defaults that drive the UI.",
        "responses": {
          "200": {
            "description": "Format configuration.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FormatsConfig" }
              }
            }
          }
        }
      }
    },
    "/download": {
      "post": {
        "operationId": "postDownload",
        "summary": "Download one audio file (single link) or a ZIP bundle (many links).",
        "description": "Pass one link to get a single 320kbps MP3 file. Pass multiple links\n(newline / comma / whitespace separated, or a public Google Sheets URL that\nexpands to many) to get a ZIP bundle of MP3s. Output is always MP3.\n\nProtected: send the caller's Firebase ID token as `Authorization: Bearer <token>`.\nAuth is enforced only when the backend is configured with Firebase credentials.",
        "security": [{ "firebaseToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/DownloadRequest" }
            },
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DownloadRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The audio file or ZIP bundle as an attachment.",
            "headers": {
              "Content-Disposition": {
                "description": "attachment; filename=\"...\" — the client derives the saved filename from this.",
                "schema": { "type": "string" }
              },
              "X-Clazzziks-Warnings": {
                "description": "Optional ` | `-separated quality warnings and per-link failures (bundle).",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "audio/mpeg": { "schema": { "type": "string", "format": "binary" } },
              "audio/wav": { "schema": { "type": "string", "format": "binary" } },
              "audio/flac": { "schema": { "type": "string", "format": "binary" } },
              "application/zip": { "schema": { "type": "string", "format": "binary" } }
            }
          },
          "400": {
            "description": "Invalid input (no links provided, or no valid links found).",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Firebase ID token.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Authenticated but not yet granted access (pending approval).",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "422": {
            "description": "The track exists but cannot be downloaded (DRM, geo-block, removed).",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "429": {
            "description": "Per-user rate limit reached (non-VIP). VIPs are exempt.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "502": {
            "description": "Unexpected download/transcode failure.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "operationId": "getMe",
        "summary": "The signed-in caller's VIP/admin status.",
        "description": "Used by the UI to show VIP state and reveal the admin link. In open/dev mode (no Firebase) the anonymous caller is reported as a local admin.",
        "security": [{ "firebaseToken": [] }],
        "responses": {
          "200": {
            "description": "Caller status.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Me" } }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Firebase ID token.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/admin/vips": {
      "get": {
        "operationId": "listVips",
        "summary": "List the VIP group (admin only).",
        "security": [{ "firebaseToken": [] }],
        "responses": {
          "200": {
            "description": "The VIP group.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VipList" } }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Firebase ID token.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Authenticated but not an admin.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "post": {
        "operationId": "addVip",
        "summary": "Add or update a VIP (admin only).",
        "security": [{ "firebaseToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VipCreate" }
            },
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/VipCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated VIP group.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VipList" } }
            }
          },
          "400": {
            "description": "Invalid input (missing/invalid email).",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Firebase ID token.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Authenticated but not an admin.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/admin/vips/{email}": {
      "patch": {
        "operationId": "updateVip",
        "summary": "Update a VIP — e.g. set their rate limit (admin only).",
        "security": [{ "firebaseToken": [] }],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VipUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated VIP group.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VipList" } }
            }
          },
          "400": {
            "description": "Invalid input (bad rate limit).",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Firebase ID token.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Authenticated but not an admin.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "No such VIP.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "delete": {
        "operationId": "removeVip",
        "summary": "Remove a VIP (admin only).",
        "security": [{ "firebaseToken": [] }],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated VIP group.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VipList" } }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Firebase ID token.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "403": {
            "description": "Authenticated but not an admin.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "No such VIP.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "firebaseToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token obtained client-side via the Firebase Auth SDK."
      }
    },
    "schemas": {
      "AudioFormat": {
        "type": "string",
        "enum": ["wav", "mp3", "flac"],
        "description": "Tolerable output formats per the project spec."
      },
      "Health": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "const": "ok" }
        }
      },
      "FormatsConfig": {
        "type": "object",
        "additionalProperties": false,
        "required": ["formats", "default_format", "bundle_format"],
        "description": "The format served to users — always MP3. Kept so the frontend can probe backend availability.",
        "properties": {
          "formats": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AudioFormat" },
            "minItems": 1
          },
          "default_format": { "$ref": "#/components/schemas/AudioFormat" },
          "bundle_format": { "$ref": "#/components/schemas/AudioFormat" }
        }
      },
      "DownloadRequest": {
        "type": "object",
        "required": ["links"],
        "description": "Output is always 320kbps MP3 — single files and ZIP bundles alike.",
        "properties": {
          "links": {
            "type": "string",
            "description": "One or more links: a single URL, many URLs (newline/comma/space separated), or a public Google Sheets URL.",
            "minLength": 1
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
      "Me": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "is_vip", "is_admin", "anonymous", "rate_limit"],
        "properties": {
          "email": { "type": ["string", "null"] },
          "is_vip": { "type": "boolean" },
          "is_admin": { "type": "boolean" },
          "anonymous": { "type": "boolean" },
          "rate_limit": {
            "type": ["integer", "null"],
            "description": "Effective tracks-per-window cap; null means unlimited."
          }
        }
      },
      "Vip": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "is_admin", "note", "rate_limit", "added_at"],
        "properties": {
          "email": { "type": "string" },
          "is_admin": { "type": "boolean" },
          "note": { "type": ["string", "null"] },
          "rate_limit": {
            "type": ["integer", "null"],
            "description": "Per-VIP tracks-per-window cap; null means unlimited."
          },
          "added_at": { "type": "string" }
        }
      },
      "VipList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["vips"],
        "properties": {
          "vips": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Vip" }
          }
        }
      },
      "VipCreate": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "minLength": 3 },
          "note": { "type": "string" },
          "is_admin": { "type": "boolean" },
          "rate_limit": {
            "type": ["integer", "null"],
            "description": "Per-VIP tracks-per-window cap; null/omitted means unlimited."
          }
        }
      },
      "VipUpdate": {
        "type": "object",
        "description": "Partial update; only provided fields change.",
        "properties": {
          "note": { "type": ["string", "null"] },
          "is_admin": { "type": "boolean" },
          "rate_limit": {
            "type": ["integer", "null"],
            "description": "Per-VIP tracks-per-window cap; null means unlimited."
          }
        }
      }
    }
  }
}